feat: confirm first if user wants to add footer
With this change, the user will get a chance to confirm if they want to
add a footer before selecting the first trailer key. The user will be
able to directly choose the trailer key for all subsequent trailers if
they want to add more trailers after entering the first trailer.

Closes: #8